+++ app/lib/analyzer/pana_runner.dart
@@ -11,7 +11,6 @@ import 'package:meta/meta.dart';
 import 'package:pana/pana.dart' hide ReportStatus;
 
 import '../job/job.dart';
-import '../package/models.dart';
 import '../package/overrides.dart';
 import '../scorecard/backend.dart';
 import '../scorecard/models.dart';
@@ -94,9 +93,10 @@ class AnalyzerJobProcessor extends JobProcessor {
         );
 
   @override
-  Future<bool> shouldProcess(PackageVersion pv, DateTime updated) {
+  Future<bool> shouldProcess(String package, String version, DateTime updated) {
     return scoreCardBackend.shouldUpdateReport(
-      pv,
+      package,
+      version,
       ReportType.pana,
       updatedAfter: updated,
     );