+++ app/lib/dartdoc/dartdoc_runner.dart
@@ -18,7 +18,6 @@ import '../frontend/static_files.dart';
 import '../job/backend.dart';
 import '../job/job.dart';
 import '../package/backend.dart';
-import '../package/models.dart';
 import '../scorecard/backend.dart';
 import '../scorecard/models.dart';
 import '../shared/configuration.dart';
@@ -170,9 +169,10 @@ class DartdocJobProcessor extends JobProcessor {
         );
 
   @override
-  Future<bool> shouldProcess(PackageVersion pv, DateTime updated) {
+  Future<bool> shouldProcess(String package, String version, DateTime updated) {
     return scoreCardBackend.shouldUpdateReport(
-      pv,
+      package,
+      version,
       ReportType.dartdoc,
       updatedAfter: updated,
     );